+++ pkg/executor/importer/table_import.go
@@ -721,7 +721,7 @@ func (ti *TableImporter) ImportSelectedRows(ctx context.Context, se sessionctx.C
 	}
 
 	return &JobImportResult{
-		Affected: uint64(dataKVCount),
+		Affected: dataKVCount,
 	}, nil
 }
 
@@ -980,7 +980,7 @@ func FlushTableStats(ctx context.Context, se sessionctx.Context, tableID int64, 
 	sessionVars := se.GetSessionVars()
 	sessionVars.TxnCtxMu.Lock()
 	defer sessionVars.TxnCtxMu.Unlock()
-	sessionVars.TxnCtx.UpdateDeltaForTable(tableID, int64(result.Affected), int64(result.Affected))
+	sessionVars.TxnCtx.UpdateDeltaForTable(tableID, result.Affected, result.Affected)
 	se.StmtCommit(ctx)
 	return se.CommitTxn(ctx)
 }
